wip: modularized indexer
