+++ apps/indexer/src/modules/auction/auction.type.ts
