+++ apps/indexer/src/modules/swaps/index.ts
@@ -1,6 +1,7 @@
 import { listenToEosTransfers } from './eos-transfers'
+import { listenToEvmTransfers } from './evm-tranfers'
 
 export function startSwapsService() {
   listenToEosTransfers()
-  listenToEosTransfers()
+  listenToEvmTransfers()
 }
